Move tests to their own file and add some
